made edit details modal compact and refactored files
