+++ src/components/Editor/EditorModals/EditWikiDetailsModal/TagsInput.tsx
@@ -1,20 +1,20 @@
 import React, { memo, useState } from 'react'
 import { Stack, Text, chakra } from '@chakra-ui/react'
 
-import { useAppDispatch, useAppSelector } from '@/store/hook'
+import { useAppDispatch } from '@/store/hook'
 import { TagsSuggestions } from '@/data/TagsSuggestions'
 import { Select, MultiValue } from 'chakra-react-select'
+import { Wiki } from '@everipedia/iq-utils'
 
 type TagValue = MultiValue<{
   label: string
   value: string
 }>
 
-const Tags = () => {
+const TagsInput = ({ wiki }: { wiki: Wiki }) => {
   const dispatch = useAppDispatch()
-  const currentWiki = useAppSelector(state => state.wiki)
   const [value, setValue] = useState<TagValue>(
-    currentWiki.tags.map(ta => ({ label: ta.id, value: ta.id })),
+    wiki.tags.map(ta => ({ label: ta.id, value: ta.id })),
   )
   const handleOnchange = (item: TagValue) => {
     setValue(item)
@@ -23,27 +23,22 @@ const Tags = () => {
       payload: item.map(ta => ({ id: ta.value })),
     })
   }
+
   return (
     <Stack spacing="4">
       <Text fontWeight="semibold">
         Tags <chakra.span opacity={0.4}>(Add up to 5)</chakra.span>
       </Text>
-      <chakra.div
-        rounded="md"
-        border="solid 1px"
-        borderColor="borderColor"
-        p={3}
-        pos="relative"
-      >
+      <chakra.div rounded="md" borderWidth={1} p={2} pos="relative">
         <Select
           isOptionDisabled={() => value.length >= 5}
           placeholder="Add tags..."
           variant="unstyled"
           isMulti
-          size="md"
+          size="sm"
           options={TagsSuggestions}
           onChange={item => handleOnchange(item)}
-          defaultValue={currentWiki.tags.map(ta => ({
+          defaultValue={wiki.tags.map(ta => ({
             label: ta.id,
             value: ta.id,
           }))}
@@ -53,4 +48,4 @@ const Tags = () => {
   )
 }
 
-export default memo(Tags)
+export default memo(TagsInput)
