+++ src/components/Editor/EditorModals/EditWikiDetailsModal/styles.ts
