+++ src/components/Editor/WikiScoreIndicator.tsx
