+++ src/pages/create-wiki/index.tsx
@@ -49,7 +49,7 @@ import { useAccount } from 'wagmi'
 import { MdTitle } from 'react-icons/md'
 import ReactCanvasConfetti from 'react-canvas-confetti'
 
-import Highlights from '@/components/Layout/Editor/Highlights/Highlights'
+import WikiDetailsSidebar from '@/components/Editor/WikiDetailsSidebar'
 import { useAppSelector } from '@/store/hook'
 import { getWikiMetadataById } from '@/utils/getWikiFields'
 import { authenticatedRoute } from '@/components/WrapperRoutes/AuthenticatedRoute'
@@ -84,7 +84,7 @@ import {
   removeDraftFromLocalStorage,
 } from '@/store/slices/wiki.slice'
 import useConfetti from '@/hooks/useConfetti'
-import WikiScoreIndicator from '@/components/Layout/Editor/WikiScoreIndicator'
+import WikiScoreIndicator from '@/components/Editor/WikiScoreIndicator'
 import { useWhiteListValidator } from '@/hooks/useWhiteListValidator'
 import { MEDIA_POST_DEFAULT_ID, WIKI_SUMMARY_LIMIT } from '@/data/Constants'
 import CreateWikiPageHeader from '@/components/SEO/CreateWikiPage'
@@ -93,7 +93,7 @@ type PageWithoutFooter = NextPage & {
   noFooter?: boolean
 }
 
-const Editor = dynamic(() => import('@/components/Layout/Editor/Editor'), {
+const Editor = dynamic(() => import('@/components/Editor/Editor'), {
   ssr: false,
 })
 
@@ -678,7 +678,7 @@ const CreateWikiContent = () => {
           <Box>
             <Skeleton isLoaded={!isLoadingWiki} w="full" h="full">
               <Center>
-                <Highlights
+                <WikiDetailsSidebar
                   initialImage={wiki?.images?.length ? wiki.images[0].id : ''}
                   isToResetImage={isNewCreateWiki}
                 />
